[OU-FIX] website: Do BS conversion with sanitization override
Certain fields had sanitize=False or other sanitization disablings (like
sanitize_attributes=False or sanitize_form=False), so when converting to
BS 5, we should allow to save the result without re-sanitizing such
content. For that, if the user is not on the sanitize overridable group,
we add it, removing it after conversion.