[OU-FIX] website: Add website editors to sanitization override group
Website editors that come from v15, where there was no sanitization,
would expect the same on the new version, or worst: they introduced some
content like iframe in 15, and re-editing that content in 16 leads to
lose it when saving it. Thus, let's add all of them to the override
sanitization group for keeping the same behavior.